+++ src/main/java/kr/codesqaud/cafe/dto/member/MemberResponseDto.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.dto;
+package kr.codesqaud.cafe.dto.member;
 
 import java.time.LocalDateTime;
