+++ src/main/java/kr/codesqaud/cafe/repository/member/MemberRepository.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.repository;
+package kr.codesqaud.cafe.repository.member;
 
 import java.util.List;
 import java.util.Optional;
